requirements.txt
 
python-telegram-bot==13.0
python-dotenv==0.15.0